refactor: fix type issues in sanity module
- Properly handle async exports in sanity-content-query.ts with 'use server' directive\n- Fixed type issues in course.ts by using proper type guards for instructor data\n- Improved collaborator.ts types with specialized interfaces for document creation\n- Removed non-async utility exports from server action file
